requirements.txt
selenium
webdriver-manager
python-docx
docx2pdf
